Refactor popover positioning logic into its own hook
